FLOW-18142: Login policy configuration
