+++ src/index.ts
@@ -79,7 +79,7 @@ export * from './service/EntityShareService';
 export * from './service/ReleaseNotesService';
 export * from './service/HistoryModuleService';
 export * from './service/AclGroupService';
-export * from './service/UserLogoutService';
+export * from './service/UserLoginService';
 export * from './service/AdminTokenService';
 export * from './service/PoiService';
 export * from './service/EntityFieldCalculationService';
