+++ src/service/UserLoginService/UserLoginPolicyController.ts
@@ -1,6 +1,8 @@
 import { APIClient, APIMapping } from '../../http';
-import { UserLogoutTypes } from './UserLogoutService.Types';
+import { UserLogoutTypes } from './UserLoginService.Types';
 import { JSONPatch } from '../DocumentTemplateService';
+
+const POLICY_PATH = '/user-logout-policy';
 
 export class UserLoginPolicyController extends APIClient {
     constructor() {
@@ -9,7 +11,7 @@ export class UserLoginPolicyController extends APIClient {
     }
 
     /**
-     * Create a Policy
+     * Create a Policy and schedule logout trigger
      * @param policy
      */
     async createPolicy(policy: UserLogoutTypes.UserLoginPolicy) {
@@ -20,7 +22,7 @@ export class UserLoginPolicyController extends APIClient {
             loginPolicy: policy.loginPolicy
         };
 
-        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>('/user-logout-policy', 'POST', createModel);
+        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>(POLICY_PATH, 'POST', createModel);
     }
 
     /**
@@ -28,7 +30,7 @@ export class UserLoginPolicyController extends APIClient {
      * @param id
      */
     async fetchPolicyById(id: string) {
-        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>(`/user-logout-policy/${id}`, 'GET');
+        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>(`${POLICY_PATH}/${id}`, 'GET');
     }
 
     /**
@@ -37,7 +39,7 @@ export class UserLoginPolicyController extends APIClient {
      * @param operations
      */
     async patchPolicy(id: string, operations: JSONPatch[]) {
-        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>(`/user-logout-policy/${id}`, 'PATCH', operations, {
+        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicy>(`${POLICY_PATH}/${id}`, 'PATCH', operations, {
             headers: { 'Content-Type': 'application/json-patch+json' },
         });
     }
@@ -47,7 +49,7 @@ export class UserLoginPolicyController extends APIClient {
      * @param id
      */
     async deletePolicy(id: string) {
-        return this.invokeApiWithErrorHandling<void>(`/user-logout-policy/${id}`, 'DELETE');
+        return this.invokeApiWithErrorHandling<void>(`${POLICY_PATH}/${id}`, 'DELETE');
     }
 
     /**
@@ -55,12 +57,7 @@ export class UserLoginPolicyController extends APIClient {
      * @param userId
      */
     async fetchPoliciesByUser(userId: string) {
-        const response = await this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicyListResponse>(`/user-logout-policy/users/${userId}`, 'GET');
-        // This is temporary and will be removed when proper endpoint will be implemented
-        if (response.isSuccessful2xx && response.data) {
-            return response.data.entries.filter((item) => Boolean(item.loginPolicy));
-        }
-        return [];
+        return this.invokeApiWithErrorHandling<UserLogoutTypes.UserLoginPolicyListResponse>(`${POLICY_PATH}/users/${userId}`, 'GET');
     }
     /**
      * Check if user can login
